+++ Assets/2DGame/Scripts/Character/CharacterScriptableObject.cs
@@ -1,5 +1,4 @@
 using UnityEngine;
-
 namespace _2DGame.Scripts.Character
 {
     [CreateAssetMenu(fileName = "Player Settings", menuName = "Settings/Player Settings ", order = 0)]
@@ -11,14 +10,12 @@ namespace _2DGame.Scripts.Character
         /// Speed of the character applied while it moves in 2D world
         /// </summary>
         public float Speed => _speed;
-    
         [Tooltip("The force applied for the character to jump")]
         [SerializeField] private float _jumpForce = 10;
         /// <summary>
         /// The force applied for the character to jump
         /// </summary>
         public float JumpForce => _jumpForce;
-    
         [Tooltip("The appearance of the character in the 2D world ")]
         [SerializeField] private Sprite characterSprite;
         /// <summary>
