+++ Assets/2DGame/Scripts/Item/SlotInventory.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 35bcf5a85fb64978810729de29e6eaf3
+timeCreated: 1675332230
